Use `EventSource` and `ChainableEvent` instead of `QuillEvent`.
