+++ local-modules/doc-client/CaretOverlay.js
@@ -4,7 +4,7 @@
 
 import { Caret } from 'doc-common';
 import { Delay } from 'promise-util';
-import { QuillEvent, QuillGeometry } from 'quill-util';
+import { QuillEvents, QuillGeometry } from 'quill-util';
 import { TObject, TString } from 'typecheck';
 
 /**
@@ -170,7 +170,7 @@ export default class CaretOverlay {
     if (false) { // eslint-disable-line no-constant-condition
       let currentEvent = this._editorComplex.bodyQuill.currentEvent;
       while (currentEvent) {
-        const selEvent = await currentEvent.nextOf(QuillEvent.SELECTION_CHANGE);
+        const selEvent = await currentEvent.nextOf(QuillEvents.SELECTION_CHANGE);
         const range    = selEvent.range;
 
         this._updateCaret(
@@ -273,11 +273,11 @@ export default class CaretOverlay {
 
     for (;;) {
       // Wait for a text change.
-      currentEvent = await currentEvent.nextOf(QuillEvent.TEXT_CHANGE);
+      currentEvent = await currentEvent.nextOf(QuillEvents.TEXT_CHANGE);
 
       // Skip any additional text changes that have already been posted, so that
       // we won't just be slowly iterating over all changes.
-      currentEvent = currentEvent.latestOfNow(QuillEvent.TEXT_CHANGE);
+      currentEvent = currentEvent.latestOfNow(QuillEvents.TEXT_CHANGE);
 
       log.detail('Got local edit event.');
       this._updateDisplay();
