+++ local-modules/quill-util/package.json
@@ -7,6 +7,7 @@
     "quill": "^1.3.1",
 
     "doc-common": "local",
+    "promise-util": "local",
     "typecheck": "local",
     "util-common": "local"
   }
